Make rem_euclid_* return a u16

+++ agb/src/display/tiled/infinite_scrolled_map.rs
@@ -63,8 +63,8 @@ impl<'a> InfiniteScrolledMap<'a> {
 
         let offset = self.current_pos - (x_start * 8, y_start * 8).into();
         let offset_scroll = (
-            self.map.size().rem_euclid_width(offset.x) as u16,
-            self.map.size().rem_euclid_height(offset.y) as u16,
+            self.map.size().rem_euclid_width(offset.x),
+            self.map.size().rem_euclid_height(offset.y),
         )
             .into();
 
@@ -179,8 +179,8 @@ impl<'a> InfiniteScrolledMap<'a> {
             self.map.set_tile(
                 vram,
                 (
-                    size.rem_euclid_width(tile_x - self.offset.x) as u16,
-                    size.rem_euclid_height(tile_y - self.offset.y) as u16,
+                    size.rem_euclid_width(tile_x - self.offset.x),
+                    size.rem_euclid_height(tile_y - self.offset.y),
                 )
                     .into(),
                 tileset,
@@ -190,8 +190,8 @@ impl<'a> InfiniteScrolledMap<'a> {
 
         let current_scroll = self.map.scroll_pos();
         let new_scroll = (
-            size.rem_euclid_width_px(current_scroll.x as i32 + difference.x) as u16,
-            size.rem_euclid_height_px(current_scroll.y as i32 + difference.y) as u16,
+            size.rem_euclid_width_px(current_scroll.x as i32 + difference.x),
+            size.rem_euclid_height_px(current_scroll.y as i32 + difference.y),
         )
             .into();
